you can now add comments

+++ server.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const app = express();
 const mysql = require('mysql');
-const bodyparser = require ('body-parser');
+const bodyparser = require('body-parser');
 app.use(express.static('public'));
 
 // Création de la connexion de mysql avec le site
@@ -46,15 +46,17 @@ app.get("/", function (req, res) {
 // Suppression des posts
 app.post("/", function (req, res) {
     console.log('del: ' + req.body.del);
-    console.log('aff:' +req.body.display);
-    let sqlDeletePost = 'DELETE FROM Post WHERE id_Post=' + req.body.del + ';';
-    connection.query(sqlDeletePost, function (error) {
-        if (error) {
-            console.log(error);
-            return;
-        }
-        res.redirect("/");
-    });
+    console.log('addcomment: ' + req.body.addcomment);
+    if (req.body.del) {
+        let sqlDeletePost = 'DELETE FROM Post WHERE id_Post=' + req.body.del + ';';
+        connection.query(sqlDeletePost, function (error) {
+            if (error) {
+                console.log(error);
+                return;
+            }
+            res.redirect("/");
+        });
+    }
 });
 
 // Définition de la route 'ajout post'
@@ -76,17 +78,49 @@ app.post("/addpost", function (req, res) {
     });
 });
 
-//Si on clique sur un post, on l'affiche dans la nouvelle page "read"
-app.get('/read/:id',function (req, res){
-    let sqlAffPost = "SELECT titre,corps,DATE_FORMAT(date_Post,'%d/%m/%Y') AS date_formated,id_Post FROM Post WHERE id_Post = "+req.params.id;
-    connection.query(sqlAffPost, function select(error, results, fields) {
+app.get("/addcomment/:id", function (req, res) {
+    console.log (req.params.id)
+    res.render("addcomment",{id:req.params.id});
+});
+
+app.post("/addcomment/:id", function (req, res) {
+    let sqlAddComm = 'INSERT INTO Commentaire (corps_Commentaire, date_Commentaire, id_Post, id_User) VALUES ("'+req.body.corps+'",NOW(),'+req.params.id+',1);'
+    connection.query(sqlAddComm, function (error, results, fields) {
         if (error) {
             console.log(error);
             return;
         }
-        if (results.length > 0) {
-            console.log(results);
-            res.render("read", {post: results});
+        res.redirect("/read/"+req.params.id);
+    });
+});
+
+//Si on clique sur un post, on l'affiche dans la nouvelle page "read"
+app.get('/read/:id', function (req, res) {
+    let sqlAffPost = "SELECT titre,corps,DATE_FORMAT(date_Post,'%d/%m/%Y') AS date_formated,id_Post FROM Post WHERE id_Post = " + req.params.id;
+    let sqlAffComm = "SELECT corps_Commentaire,DATE_FORMAT(date_Commentaire,'%d/%m/%Y') AS datec_formated, id_Commentaire FROM Commentaire WHERE id_Post = " + req.params.id;
+    connection.query(sqlAffPost, function select(error, resultp, fields) {
+        if (error) {
+            console.log(error);
+            return;
+        }
+        if (resultp.length > 0) {
+            connection.query(sqlAffComm, function select(error, resultc, fields) {
+                if (error) {
+                    console.log(error);
+                    return;
+                }
+                if (resultc.length > 0) {
+                    res.render("read", {
+                        post: resultp,
+                        comments: resultc
+                    });
+                    return (resultc);
+                } else {
+                    res.render("read", {
+                        post: resultp
+                    });
+                }
+            });
         } else {
             console.log("Pas de données");
             res.render('index');
